+++ Assets/Script/MeshDisplayInfo.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7263dc88570d2ae47bc15b46780486cd
+guid: 6e5fd8e7ec40cf543a5378ec79dd916e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
