--- Assets/Script/MeshDisplayInfo2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6e5fd8e7ec40cf543a5378ec79dd916e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
