--- Assets/Script/MeshGenerator2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9378f370d1eec5d48a824f53da6d3e09
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
